+++ Sources/TMDb/Domain/Models/CreditTVSeries.swift
@@ -55,6 +55,8 @@ Hashable, Sendable {
     ///
     /// TV series first air date.
     ///
+    /// Midnight GMT on the day TMDb reports.
+    ///
     public let firstAirDate: Date?
 
     ///