+++ Sources/TMDb/Domain/Models/MovieCastCredit.swift
@@ -48,6 +48,8 @@ public struct MovieCastCredit: Identifiable, Codable, Equatable, Hashable, Senda
     ///
     /// Movie release date.
     ///
+    /// Midnight GMT on the day TMDb reports.
+    ///
     public let releaseDate: Date?
 
     ///